Release v5.5

+++ include/views/overlays/fireworks.h
@@ -30,7 +30,7 @@ private:
         ParticleType type : 3;
         uint8_t phase : 5;
         int8_t y;
-        int8_t x;
+        int16_t x;
         int8_t vy;
         int8_t vx;
     };
